# openapi.yaml
openapi: "3.0.3"
info:
  title: GraphOne API
  description: |
    Intelligence layer for the AI economy — companies, investors, founders, products, funding rounds, and news.
    
    **Authentication**: Write operations (POST, PATCH, DELETE) require an `X-API-Key` header.
    **Rate Limiting**: 100 requests per minute per IP.
    **Response shape**: All responses follow `{ data, meta }` on success or `{ error: { code, message } }` on failure.
  version: "1.0.0"
  contact:
    name: GraphOne Engineering
servers:
  - url: http://localhost:4000
    description: Local development
  - url: https://your-deployed-url.railway.app
    description: Production
components:
  securitySchemes:
    ApiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
...
paths:
...
  /api/companies/{slug}/funding:
    get:
      summary: Funding timeline for a company
      parameters:
        - in: path
          name: slug
          required: true
          schema: { type: string }
      responses:
        "200":
          description: Funding rounds ordered chronologically
  /api/companies/{slug}/products:
    get:
      summary: Products by company
      parameters:
        - in: path
          name: slug
          required: true
          schema: { type: string }
      responses:
        "200":
          description: Product list
  /api/companies/{slug}/graph:
    get:
      summary: 1-hop ecosystem graph (investors, products, competitors)
      parameters:
        - in: path
          name: slug
          required: true
          schema: { type: string }
      responses:
        "200":
          description: Graph data
...
  /api/investors/{slug}/investments:
    get:
      summary: Paginated investment history
      parameters:
        - in: path
          name: slug
          required: true
          schema: { type: string }
        - in: query
          name: page
          schema: { type: integer }
      responses:
        "200":
          description: Investment list
  /api/investors/{slug}/co-investors:
    get:
      summary: Who this investor most often syndicates with
      parameters:
        - in: path
          name: slug
          required: true
          schema: { type: string }
      responses:
        "200":
          description: Co-investor frequency list
...
  /api/products:
    get:
      summary: List products (filter by category, sort by popular/newest)
      responses:
        "200":
          description: Product list
    post:
      summary: Create product
      security: [{ ApiKeyAuth: [] }]
      responses:
        "201":
          description: Created
  /api/products/{slug}:
    get:
      summary: Product detail
      parameters:
        - in: path
          name: slug
          required: true
          schema: { type: string }
      responses:
        "200":
          description: Product
  /api/news:
    get:
      summary: Paginated news feed with optional tag filter
      parameters:
        - in: query
          name: tag
          schema: { type: string }
        - in: query
          name: page
          schema: { type: integer }
      responses:
        "200":
          description: News articles
    post:
      summary: Create news article
      security: [{ ApiKeyAuth: [] }]
      responses:
        "201":
          description: Created
...